+++ .gitignore
@@ -1,3 +1,2 @@
 .DS_Store
 node_modules/
-data/
