+++ Dockerfile-dev
@@ -16,9 +16,5 @@ RUN npm install
 # Bundle app source
 COPY . .
 
-## Wait mongoDB to be accepting connections
-ADD https://github.com/ufoscout/docker-compose-wait/releases/download/2.4.0/wait /wait
-RUN chmod +x /wait
-
 EXPOSE 8081
-CMD /wait && npm run devStart
+CMD [ "npm", "run", "devStart" ]
