--- app/database_config.js
@@ -1,3 +1,0 @@
-module.exports = {
-    'url' : 'mongodb://mongo:27017'
-};
