+++ docker-compose.yml
@@ -10,12 +10,3 @@ services:
      - .:/code
     environment:
       - NODE_ENV=production
-    depends_on:
-      - "mongo"
-  mongo:
-      container_name: mongo
-      image: mongo
-      volumes:
-        - ./data:/data/db
-      ports:
-        - "27017:27017"
