+++ package.json
@@ -22,8 +22,7 @@
     "babel-cli": "^6.26.0",
     "babel-preset-env": "^1.7.0",
     "body-parser": "^1.18.3",
-    "express": "^4.16.4",
-    "mongoose": "^5.3.6"
+    "express": "^4.16.4"
   },
   "devDependencies": {
     "nodemon": "^1.18.4"
